extract device pin function to separate file
